Get rid of most uses of Float and instead use Int to prevent floating point math errors and fractional amounts of cargo being left around. Fixed issue with traders buying goods for more than the value of the known delivery price, and added a master negotiator trait that can get a discount. Speed up the simulation, and added a bunch more goods.

+++ src/main/kotlin/com/ostaragame/systems/economy/actors/Cargo.kt
@@ -4,4 +4,4 @@ import com.ostaragame.systems.economy.engine.TradeGood
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Cargo (val tradeGood: TradeGood, var units:Float)
+data class Cargo (val tradeGood: TradeGood, var units:Int)
